+++ sirbot/core/__init__.py
@@ -1,5 +1,4 @@
-from .core import SirBot  # noQa
+from .core import SirBot, registry  # noQa
 from .__meta__ import DATA as METADATA  # noQa
 from .hookimpl import hookimpl  # noQa
 from .plugin import Plugin  # noQa
-from .registry import registry  # noQa
