+++ sirbot/core/core.py
@@ -14,15 +14,16 @@ import aiohttp
 import pluggy
 import yaml
 
-from collections import defaultdict
+from collections import defaultdict, MutableMapping
 from aiohttp import web
 
 from sirbot.utils import merge_dict
 
 from . import hookspecs
-from .registry import registry
+from .errors import FrozenRegistryError
 
 logger = logging.getLogger(__name__)
+registry = None
 
 if sys.version_info[:2] == (3, 5):
     ModuleNotFoundError = ImportError
@@ -38,6 +39,8 @@ class SirBot:
         loop (asyncio.AbstractEventLoop): Event loop
     """
     def __init__(self, config=None, *, loop=None):
+        global registry
+        registry = RegistrySingleton()
         self.config = config or {}
         self._configure()
         logger.info('Initializing Sir Bot-a-lot')
@@ -256,3 +259,51 @@ class SirBot:
         """
         self._loop.run_until_complete(self._configure_plugins())
         web.run_app(self._app, host=host, port=port)  # pragma: no cover
+
+
+class RegistrySingleton(MutableMapping):
+    """
+    Class regrouping all available plugin factory.
+
+    The plugin factory is called each time the registry is queried.
+    It is good practice to keep interacting on the same factory result without
+    querying the registry each time.
+
+    Similar behaviour as a dictionary
+    """
+    def __init__(self):
+        super().__init__()
+        self._frozen = False
+        self._plugins = dict()
+
+    @property
+    def frozen(self):
+        return self._frozen
+
+    @frozen.setter
+    def frozen(self, _):
+        raise ValueError('Read only property')
+
+    def freeze(self):
+        self._frozen = True
+
+    def __getitem__(self, item):
+        return self._plugins[item]()
+
+    def __setitem__(self, key, value):
+        if not self._frozen:
+            self._plugins[key] = value
+        else:
+            raise FrozenRegistryError()
+
+    def __delitem__(self, key):
+        raise FrozenRegistryError()
+
+    def __iter__(self):
+        return iter(self._plugins)
+
+    def __len__(self):
+        return len(self._plugins)
+
+    def __contains__(self, item):
+        return item in self._plugins
